GH-78: Создать интерфейсы для controller-ов и service-ов и переделать готовые классы в implementation-ы

+++ api/src/main/java/ru/teamtwo/api/repository/customer/CartItemRepository.java
@@ -5,6 +5,6 @@ import ru.teamtwo.api.models.customer.CartItem;
 
 import java.util.Set;
 
-public interface CartItemRepository  extends JpaRepository<CartItem, Integer> {
-    Set<CartItem> getCartItemsByCustomer_Id(Integer customer_id);
+public interface CartItemRepository  extends JpaRepository<CartItem, Long> {
+    Set<CartItem> getCartItemsByCustomer_Id(Long customer_id);
 }
